+++ gridstatus/nyiso.py
@@ -633,6 +633,36 @@ class NYISO(ISOBase):
 
         return df
 
+    def get_capacity_prices(self, date=None, verbose=False):
+        """Pull the most recent capacity market report's market clearing prices
+
+        Parameters:
+            date (pd.Timestamp): date that will be used to pull latest capacity report (will refer to month and year)
+            verbose (bool): print out requested url
+
+        Returns:
+            pd.DataFrame: a dataframe of monthly capacity prices (all three auctions) for each of the four capacity localities within NYISO
+        """
+        if date is None:
+            date = pd.Timestamp.now(tz=self.default_timezone)
+        else:
+            date = utils._handle_date(date, tz=self.default_timezone)
+
+        # todo: it looks like the "27447313" component of the base URL changes every year but I'm not sure what the link between that and the year is...
+        capacity_market_base_url = (
+            "https://www.nyiso.com/documents/20142/27447313/ICAP-Market-Report"
+        )
+
+        url = f"{capacity_market_base_url}-{date.month_name()}-{date.year}.xlsx"
+        if verbose:
+            print(url)
+
+        df = pd.read_excel(url, sheet_name="MCP Table", header=[0, 1])
+
+        df.rename(columns={"Unnamed: 0_level_0": "", "Date": ""}, inplace=True)
+        df.set_index("", inplace=True)
+        return df.dropna(how="any", axis="columns")
+
 
 def _handle_time(df):
     if "Time Stamp" in df.columns: